backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pgvector==0.2.4

# AI & Embeddings
openai==1.10.0
tavily-python==0.3.0

# Authentication
fastapi-users[sqlalchemy]==12.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# WebSocket
websockets==12.0

# Document Processing (for future phases)
pypdf2==3.0.1
pdfplumber==0.10.3
pillow==10.2.0
pytesseract==0.3.10
reportlab==4.0.9

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.2
httpx==0.27.0
